feat: resplit index.ts for gpio
